Adding new common data a fake class files
Test files share common data an mock classes
